Split classes in different documents
